필요없는 import 제거, lifecycle-extenstions version-up
사진 첨부하는 것 해야 됨.

+++ app/src/main/java/com/whalez/programmerslineplus/ui/home/MainActivity.kt
@@ -1,33 +1,30 @@
 package com.whalez.programmerslineplus.ui.home
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.view.ContextThemeWrapper
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
-import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import com.skydoves.powermenu.kotlin.powerMenu
+import com.whalez.programmerslineplus.R
+import com.whalez.programmerslineplus.data.Memo
+import com.whalez.programmerslineplus.ui.edit.EditMemoActivity
+import com.whalez.programmerslineplus.ui.home.menu.MenuFactory
+import com.whalez.programmerslineplus.ui.home.menu.MenuFactory.Companion.APP_INFO
+import com.whalez.programmerslineplus.ui.home.menu.MenuFactory.Companion.DELETE_ALL
 import com.whalez.programmerslineplus.utils.ConstValues.Companion.ADD_MEMO_REQUEST
 import com.whalez.programmerslineplus.utils.ConstValues.Companion.EDIT_MEMO_REQUEST
 import com.whalez.programmerslineplus.utils.ConstValues.Companion.EXTRA_CONTENT
 import com.whalez.programmerslineplus.utils.ConstValues.Companion.EXTRA_ID
 import com.whalez.programmerslineplus.utils.ConstValues.Companion.EXTRA_TITLE
-import com.whalez.programmerslineplus.ui.home.menu.MenuFactory
-import com.whalez.programmerslineplus.ui.home.menu.MenuFactory.Companion.APP_INFO
-import com.whalez.programmerslineplus.ui.home.menu.MenuFactory.Companion.DELETE_ALL
-import com.whalez.programmerslineplus.R
-import com.whalez.programmerslineplus.data.Memo
-import com.whalez.programmerslineplus.ui.edit.EditMemoActivity
 import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 
 class MainActivity : AppCompatActivity() {
 
@@ -47,7 +44,7 @@ class MainActivity : AppCompatActivity() {
         val memoAdapter = MemoAdapter()
         rv_main.adapter = memoAdapter
 
-        memoViewModel = ViewModelProviders.of(this)[MemoViewModel::class.java]
+        memoViewModel = ViewModelProvider(this).get(MemoViewModel::class.java)
         memoViewModel.getAll().observe(this,
             Observer<List<Memo>> { memos -> memoAdapter.submitList(memos) })
 
@@ -57,13 +54,24 @@ class MainActivity : AppCompatActivity() {
             startActivityForResult(intent, ADD_MEMO_REQUEST)
         }
 
+        // 메뉴 버튼 클릭
         btn_menu.setOnClickListener { mainMenu.showAsDropDown(it) }
         mainMenu.setOnMenuItemClickListener { position, item ->
             when (position) {
                 DELETE_ALL -> {
-                    // 쿼리문 비동기 처리
-                    memoViewModel.deleteAll()
-                    Toast.makeText(this, "모두 삭제되었습니다.", Toast.LENGTH_SHORT).show()
+                    val builder = AlertDialog.Builder(
+                        ContextThemeWrapper(
+                            this@MainActivity,
+                            R.style.MyAlertDialogStyle
+                        )
+                    )
+                    builder.setMessage("모든 메모를 삭제하시겠습니까?")
+                        .setPositiveButton("예") { _, _ ->
+                            memoViewModel.deleteAll()
+                            Toast.makeText(this@MainActivity, "모든 메모가 삭제되었습니다.", Toast.LENGTH_SHORT).show()
+                        }
+                        .setNegativeButton("아니요") { _, _ -> }
+                        .show()
                 }
                 APP_INFO -> {
                     Toast.makeText(this, item.title, Toast.LENGTH_SHORT).show()
@@ -72,16 +80,17 @@ class MainActivity : AppCompatActivity() {
             }
         }
 
-        ItemTouchHelper(object: ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
+        ItemTouchHelper(object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
+
             override fun onMove(
                 recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder,
-                target: RecyclerView.ViewHolder
+                viewHolder: ViewHolder,
+                target: ViewHolder
             ): Boolean {
                 return false
             }
 
-            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
+            override fun onSwiped(viewHolder: ViewHolder, direction: Int) {
                 val builder = AlertDialog.Builder(
                     ContextThemeWrapper(
                         this@MainActivity,
@@ -91,38 +100,32 @@ class MainActivity : AppCompatActivity() {
                 builder.setMessage("정말 삭제하시겠습니까?")
                     .setPositiveButton("예") { _, _ ->
                         memoViewModel.delete(memoAdapter.getMemoAt(viewHolder.adapterPosition))
-                        Toast.makeText(this@MainActivity, "삭제 완료", Toast.LENGTH_SHORT).show()
+                        Toast.makeText(this@MainActivity, "삭제되었습니다.", Toast.LENGTH_SHORT).show()
                     }
-                    .setNegativeButton("아니요") {_, _ ->
+                    .setNegativeButton("아니요") { _, _ ->
                         memoAdapter.notifyDataSetChanged()
                     }
                     .show()
             }
         }).attachToRecyclerView(rv_main)
 
-        memoAdapter.setOnItemClickListener(object:
+        // 아이템 클릭
+        memoAdapter.setOnItemClickListener(object :
             MemoAdapter.OnItemClickListener {
             override fun onItemClick(memo: Memo) {
-                Log.d("kkk", "itemClicked!!!!!!!!!!!!")
                 val intent = Intent(this@MainActivity, EditMemoActivity::class.java)
                 intent.putExtra(EXTRA_ID, memo.id)
                 intent.putExtra(EXTRA_TITLE, memo.title)
                 intent.putExtra(EXTRA_CONTENT, memo.content)
-//                startActivity(intent)
                 startActivityForResult(intent, EDIT_MEMO_REQUEST)
-
             }
-
         })
-
-
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
 
-        if(requestCode == ADD_MEMO_REQUEST && resultCode == RESULT_OK) {
-//            Log.d("kkk", data.getStringExtra(EXTRA_CONTENT))
+        if (requestCode == ADD_MEMO_REQUEST && resultCode == RESULT_OK) {
             val title = data!!.getStringExtra(EXTRA_TITLE)
             val content = data.getStringExtra(EXTRA_CONTENT)
             val memo = Memo(title!!, content!!)
@@ -130,8 +133,7 @@ class MainActivity : AppCompatActivity() {
             memoViewModel.insert(memo)
 
             Toast.makeText(this, "메모 저장완료", Toast.LENGTH_SHORT).show()
-        }
-        else if(requestCode == EDIT_MEMO_REQUEST && resultCode == RESULT_OK) {
+        } else if (requestCode == EDIT_MEMO_REQUEST && resultCode == RESULT_OK) {
             val id = data!!.getIntExtra(EXTRA_ID, -1)
             if (id == -1) {
                 Toast.makeText(this, "메모가 수정되지 않았습니다!", Toast.LENGTH_SHORT).show()
@@ -145,12 +147,9 @@ class MainActivity : AppCompatActivity() {
 
             memoViewModel.update(memo)
 
-            Toast.makeText(this, "새 메모가 저장되었습니다!", Toast.LENGTH_SHORT).show()
-        }
-        else {
+            Toast.makeText(this, "새 메모가 수정되었습니다.", Toast.LENGTH_SHORT).show()
+        } else {
             Toast.makeText(this, "새 메모가 저장되지 않았습니다!", Toast.LENGTH_SHORT).show()
-
         }
     }
-
 }
